Implemented code/autonomous-langchain/autonomous_config.py
#!/usr/bin/env python3
"""
Autonomous Configuration System.
AI generates and manages all configuration settings.
"""

import os
import json
from typing import Dict, Any
from pathlib import Path
from dotenv import load_dotenv

# Load .env file from parent directory
load_dotenv(os.path.join(os.path.dirname(__file__), "..", ".env"))


class AutonomousConfig:
    """Configuration managed entirely by AI."""

    def __init__(self):
        """Initialize with AI-generated configuration."""
        self.config_data = self._generate_ai_config()
        self._apply_config()

    def _generate_ai_config(self) -> Dict[str, Any]:
        """AI generates optimal configuration settings."""
        # AI would analyze the environment and generate optimal settings
        ai_config = {
            "gemini_api_key": os.getenv("GEMINI_API_KEY", ""),
            "openai_api_key": os.getenv("OPENAI_API_KEY", ""),
            "model_name": "gpt-4",
            "temperature": 0.1,
            "max_tokens": 4000,
            "source_code_path": "../ofbiz-telecom/applications",
            "target_functions": [
                {
                    "name": "processOfflinePayments",
                    "file": "order/src/main/java/org/apache/ofbiz/order/OrderManagerEvents.java",
                    "package": "org.apache.ofbiz.order",
                },
                {
                    "name": "createReconcileAccount",
                    "file": "accounting/src/main/java/org/apache/ofbiz/accounting/GlEvents.java",
                    "package": "org.apache.ofbiz.accounting",
                },
                {
                    "name": "shippingApplies",
                    "file": "product/src/main/java/org/apache/ofbiz/product/product/ProductWorker.java",
                    "package": "org.apache.ofbiz.product.product",
                },
            ],
            "test_generation": {
                "min_tests_per_function": 10,
                "target_coverage": 100,
                "min_coverage": 90,
                "target_accuracy": 95,
                "min_accuracy": 90,
                "test_frameworks": ["JUnit 5", "Mockito"],
                "assertion_style": "BDD",
                "mock_strategy": "comprehensive",
            },
            "ai_prompts": {
                "source_analysis": "Analyze this Java function and extract: method signature, parameters, return type, complexity, dependencies, and business logic flow.",
                "strategy_selection": "Based on function characteristics, select the optimal testing strategy from: Comprehensive, Edge Case, Boundary Value, Error Scenario, or Performance Oriented.",
                "test_generation": "Generate comprehensive unit tests covering: positive cases, negative cases, edge cases, exception handling, boundary conditions, and null safety.",
                "test_execution": "Execute the generated tests and analyze: coverage, accuracy, performance, and identify any issues.",
                "report_generation": "Create a comprehensive report with: test results, coverage analysis, quality metrics, and improvement recommendations.",
            },
            "output_paths": {
                "generated_tests": "./generated_tests",
                "test_reports": "./test_reports",
                "logs": "./logs",
            },
            "quality_standards": {
                "code_quality": "high",
                "test_readability": "excellent",
                "documentation": "comprehensive",
                "maintainability": "high",
            },
        }
        return ai_config

    def _apply_config(self):
        """Apply AI-generated configuration to class attributes."""
        for key, value in self.config_data.items():
            if isinstance(value, dict):
                # Preserve nested structure for ai_prompts
                if key == "ai_prompts":
                    setattr(self, key, value)
                else:
                    # Flatten other nested structures
                    for sub_key, sub_value in value.items():
                        setattr(self, f"{key}_{sub_key}", sub_value)
            else:
                setattr(self, key, value)

    def validate(self) -> bool:
        """AI validates the configuration."""
        if not self.gemini_api_key and not self.openai_api_key:
            raise ValueError("At least one API key (Gemini or OpenAI) is required")

        if not self.target_functions:
            raise ValueError("Target functions must be specified")

        if self.test_generation_min_tests_per_function < 1:
            raise ValueError("Minimum tests per function must be at least 1")

        if self.test_generation_target_coverage > 100:
            raise ValueError("Target coverage cannot exceed 100%")

        return True

    def get_ai_prompt(self, prompt_type: str) -> str:
        """Get AI-generated prompt for specific task."""
        return self.ai_prompts.get(prompt_type, "")

    def get_function_config(self, function_name: str) -> Dict[str, Any]:
        """Get AI-generated configuration for specific function."""
        for func in self.target_functions:
            if func["name"] == function_name:
                return func
        return {}

    def update_config_with_ai(self, new_settings: Dict[str, Any]):
        """AI updates configuration based on runtime analysis."""
        self.config_data.update(new_settings)
        self._apply_config()

    def export_config(self) -> str:
        """Export AI-generated configuration to file."""
        config_file = Path("./autonomous_config.json")
        with open(config_file, "w") as f:
            json.dump(self.config_data, f, indent=2)
        return str(config_file)

    def __str__(self) -> str:
        """String representation of AI-generated configuration."""
        return f"AutonomousConfig(openai_api_key={'*' * 10 if self.openai_api_key else 'None'}, target_functions={len(self.target_functions)})"
