Implemented code/autonomous-langchain/requirements.txt
langchain-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
